@@ -1,8 +1,7 @@
 {
   "name": "gatsby-theme-blog-ocean",
-  "version": "0.1.5",
+  "version": "1.0.0",
   "description": "A blog theme for Gatsby.",
-  "main": "gatsby-config.js",
   "repository": "https://github.com/HowyoungZhou/gatsby-theme-blog-ocean.git",
   "author": "Howyoung <howyoung3@gmail.com>",
   "license": "MIT",
@@ -22,48 +21,58 @@
     "extract": "yarn run babel --config-file ./babel-extract.config.js -o tmp/chunk.js 'src/**/*.{js,jsx,ts,tsx}' && rm -rf tmp"
   },
   "peerDependencies": {
-    "gatsby": "^4.4.0"
+    "gatsby": "^5.11.0"
   },
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@iconify-icons/emojione": "^1.1.1",
-    "@iconify-icons/logos": "^1.1.14",
-    "@iconify/react": "^3.1.3",
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "@mui/icons-material": "^5.2.4",
-    "@mui/lab": "^5.0.0-alpha.62",
-    "@mui/material": "^5.2.4",
-    "gatsby-core-utils": "^3.5.2",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@giscus/react": "^2.3.0",
+    "@iconify-icons/emojione": "^1.2.6",
+    "@iconify-icons/logos": "^1.2.33",
+    "@iconify/react": "^4.1.1",
+    "@mdx-js/mdx": "^2.3.0",
+    "@mdx-js/react": "^2.3.0",
+    "@mui/icons-material": "^5.14.3",
+    "@mui/lab": "^5.0.0-alpha.140",
+    "@mui/material": "^5.14.5",
+    "gatsby-core-utils": "^4.11.0",
+    "gatsby-material-ui-components": "^5.3.0",
     "gatsby-plugin-feed-mdx": "^1.1.0",
-    "gatsby-plugin-image": "^2.4.0",
-    "gatsby-plugin-manifest": "^4.6.0",
-    "gatsby-plugin-mdx": "^3.4.0",
-    "gatsby-plugin-offline": "^5.6.0",
-    "gatsby-plugin-react-helmet": "^5.5.0",
-    "gatsby-plugin-react-i18next": "^1.2.2",
-    "gatsby-plugin-sharp": "^4.4.0",
-    "gatsby-plugin-typescript": "^4.4.0",
-    "gatsby-remark-copy-linked-files": "^5.7.0",
-    "gatsby-remark-images": "^6.7.0",
-    "gatsby-remark-smartypants": "^5.7.0",
-    "gatsby-source-filesystem": "^4.4.0",
-    "gatsby-theme-material-ui": "^5.1.0",
-    "gatsby-transformer-sharp": "^4.4.0",
-    "i18next": "^21.6.3",
-    "path-to-regexp": "^6.2.0",
-    "prism-react-renderer": "^1.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "gatsby-plugin-image": "^3.11.0",
+    "gatsby-plugin-manifest": "^5.11.0",
+    "gatsby-plugin-material-ui": "^4.1.0",
+    "gatsby-plugin-mdx": "^5.11.0",
+    "gatsby-plugin-offline": "^6.11.0",
+    "gatsby-plugin-react-helmet": "^6.11.0",
+    "gatsby-plugin-react-i18next": "^3.0.1",
+    "gatsby-plugin-sharp": "^5.11.0",
+    "gatsby-plugin-typescript": "^5.11.0",
+    "gatsby-plugin-webfonts": "^2.3.2",
+    "gatsby-remark-copy-linked-files": "^6.11.0",
+    "gatsby-remark-images": "^7.11.0",
+    "gatsby-remark-smartypants": "^6.11.0",
+    "gatsby-source-filesystem": "^5.11.0",
+    "gatsby-transformer-sharp": "^5.11.0",
+    "i18next": "^23.4.5",
+    "path-to-regexp": "^6.2.1",
+    "prism-react-renderer": "^2.0.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
-    "react-i18next": "^11.15.1",
-    "remark-slug": "6.1.0"
+    "react-i18next": "^13.2.0",
+    "rehype-mathjax": "^4.0.3",
+    "rehype-slug": "^5.1.0",
+    "remark-gfm": "^3.0.1",
+    "remark-math": "^5.1.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.16.8",
-    "@babel/plugin-transform-typescript": "^7.16.8",
-    "@types/react-helmet": "^6.1.5",
-    "babel-plugin-i18next-extract": "^0.8.3"
+    "@babel/cli": "^7.22.10",
+    "@babel/plugin-transform-typescript": "^7.22.10",
+    "@types/node": "^20.5.3",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-helmet": "^6.1.6",
+    "babel-plugin-i18next-extract": "^0.9.0",
+    "typescript": "^5.1.6"
   }
 }
