@@ -1,13 +1,12 @@
-const { match, compile } = require("path-to-regexp");
-const path = require(`path`)
-const {
+import { match, compile } from "path-to-regexp";
+import path from "path";
+import {
   createFilePath,
   createRemoteFileNode,
-} = require(`gatsby-source-filesystem`)
-const { createContentDigest, slash } = require(`gatsby-core-utils`)
+} from "gatsby-source-filesystem";
 
 
-exports.pluginOptionsSchema = ({ Joi }) => {
+export const pluginOptionsSchema = ({ Joi }) => {
   return Joi.object({
     postsSource: Joi.string().default('posts').description('Name of the posts source passed to the `gatsby-source-filesystem` plugin'),
     postsFolderLayout: Joi.string().default('/:lang/:segment+').description('Layout of the posts folder'),
@@ -24,6 +23,7 @@ exports.pluginOptionsSchema = ({ Joi }) => {
     rssTitle: Joi.string().description('Title of the RSS feed'),
     maxImageWidth: Joi.number().min(0).default(1920).description('Max width of images generated by the `gatsby-plugin-sharp` plugin'),
     manifest: Joi.object().default({}).description('Custom PWA manifest'),
+    giscusOptions: Joi.object().default({}).description('Custom Giscus options'),
   });
 }
 
@@ -45,6 +45,15 @@ const mdxResolverPassthrough = (fieldName) => async (
   return result;
 }
 
+// from gatsby-core-utils, as it can not be used due to incompatibility with ESM
+function slash(path) {
+  const isExtendedLengthPath = /^\\\\\?\\/.test(path);
+  if (isExtendedLengthPath) {
+    return path;
+  }
+  return path.replace(/\\/g, `/`);
+}
+
 async function processRelativeImage(source, context, type) {
   // Image is a relative path - find a corresponding file
   const mdxFileNode = context.nodeModel.findRootNodeAncestor(
@@ -52,7 +61,7 @@ async function processRelativeImage(source, context, type) {
     (node) => node.internal && node.internal.type === `File`
   )
   if (!mdxFileNode) return;
-  const imagePath = slash(path.join(mdxFileNode.dir, source[type]))
+  const imagePath = slash(path.join(mdxFileNode.dir, source[type]));
 
   return await context.nodeModel.findOne({
     type: 'File',
@@ -64,14 +73,13 @@ async function processRelativeImage(source, context, type) {
   });
 }
 
-exports.createSchemaCustomization = ({ actions, schema }, themeOptions) => {
+export const createSchemaCustomization = ({ actions, schema }, themeOptions) => {
   const { excerptLength } = themeOptions;
   const { createTypes } = actions;
   createTypes(`interface BlogPost implements Node {
       id: ID!
       title: String!
       body: String!
-      html: String!
       slug: String!
       date: Date! @dateformat
       tags: [String]!
@@ -146,10 +154,6 @@ exports.createSchemaCustomization = ({ actions, schema }, themeOptions) => {
           type: `String!`,
           resolve: mdxResolverPassthrough(`body`),
         },
-        html: {
-          type: `String!`,
-          resolve: mdxResolverPassthrough(`html`),
-        },
       },
       interfaces: [`Node`, `BlogPost`],
       extensions: {
@@ -169,8 +173,8 @@ function validURL(str) {
 }
 
 // Create fields for post slugs and source
-exports.onCreateNode = async (
-  { node, actions, getNode, createNodeId, store, cache },
+export const onCreateNode = async (
+  { node, actions, getNode, createNodeId, store, cache,createContentDigest },
   themeOptions
 ) => {
   const { createNode, createParentChildLink } = actions;
@@ -268,17 +272,25 @@ exports.onCreateNode = async (
   createParentChildLink({ parent: node, child: getNode(mdxBlogPostId) });
 }
 
+import { createRequire } from "module";
+const require = createRequire(import.meta.url);
+
 const PostTemplate = require.resolve(`./src/templates/post-query`);
 
-exports.createPages = async ({ graphql, actions, reporter }) => {
+export const createPages = async ({ graphql, actions, reporter }) => {
   const { createPage } = actions;
 
   const result = await graphql(`
     {
-      allBlogPost(sort: { fields: [date, title], order: DESC }, limit: 1000) {
+      allBlogPost(sort: [{date: DESC}, {title: ASC}], limit: 1000) {
         nodes {
           id
           slug
+          parent {
+            internal {
+              contentFilePath
+            }
+          }
         }
       }
     }
@@ -296,10 +308,9 @@ exports.createPages = async ({ graphql, actions, reporter }) => {
   posts.forEach((post, index) => {
     const previous = index === posts.length - 1 ? null : posts[index + 1];
     const next = index === 0 ? null : posts[index - 1];
-    const { slug } = post;
     createPage({
-      path: slug,
-      component: PostTemplate,
+      path: post.slug,
+      component: `${PostTemplate}?__contentFilePath=${post.parent.internal.contentFilePath}`,
       context: {
         id: post.id,
         previousId: previous ? previous.id : undefined,
