@@ -1,7 +1,8 @@
 import React from "react";
 import { useStaticQuery, graphql } from "gatsby";
 import { useTheme } from "@mui/material";
-import { Helmet } from "gatsby-plugin-react-i18next";
+import {Helmet as ReactHelmet, HelmetProps} from 'react-helmet';
+import useI18n from '../utils/use-i18n';
 
 type MetaProps = JSX.IntrinsicElements["meta"];
 
@@ -11,6 +12,28 @@ export declare interface SeoProps {
   meta?: MetaProps[];
   title?: string;
 }
+
+// taken from https://github.com/microapps/gatsby-plugin-react-i18next/blob/v2.0.5/src/Helmet.tsx
+// will migrate to Head API when it supports i18n context
+export const Helmet: React.FC<HelmetProps> = ({children, ...props}) => {
+  const {languages, language, originalPath, defaultLanguage, siteUrl = ''} = useI18n();
+  const createUrlWithLang = (lng: string) => {
+    const url = `${siteUrl}${lng === defaultLanguage ? '' : `/${lng}`}${originalPath}`;
+    return url.endsWith('/') ? url : `${url}/`;
+  };
+  return (
+    <ReactHelmet {...props}>
+      <html lang={language} />
+      <link rel="canonical" href={createUrlWithLang(language)} />
+      {languages.map((lng) => (
+        <link rel="alternate" key={lng} href={createUrlWithLang(lng)} hrefLang={lng} />
+      ))}
+      {/* adding a fallback page for unmatched languages */}
+      <link rel="alternate" href={createUrlWithLang(defaultLanguage)} hrefLang="x-default" />
+      {children}
+    </ReactHelmet>
+  );
+};
 
 export default function Seo({ description, meta = [], title }: SeoProps) {
   const { site } = useStaticQuery(
@@ -70,11 +93,13 @@ export default function Seo({ description, meta = [], title }: SeoProps) {
     }
   ];
 
+  const metaProps = [...defaultMeta, ...meta];
   return (
-    <Helmet
-      title={title}
-      titleTemplate={defaultTitle ? `%s | ${defaultTitle}` : null}
-      meta={defaultMeta.concat(meta)}
-    />
+    <Helmet defer={false}>
+      <title>{defaultTitle ? `${title} | ${defaultTitle}` : title}</title>
+      <meta name="viewport" content="initial-scale=1, width=device-width" />
+      {metaProps.map((props, i) => <meta key={i} {...props} />)}
+       {/* TODO: lang link when use Head API */}
+    </Helmet>
   );
 }
